fix: linting errors in ItemContext.js file, data set errors in home page and large additions to the item page which is not completed
