+++ src/components/SliderCard/index.jsx
@@ -3,9 +3,9 @@ import { Link } from "react-router-dom";
 import PropTypes from "prop-types";
 import styles from "./SliderCard.module.css";
 
-function SliderCard({ title, desc, price, src, link }) {
+function SliderCard({ title, desc, price, src }) {
   return (
-    <Link to="/item" state={{ title, desc, price, src, link }}>
+    <Link to="/item" state={{ title, desc, price, src }}>
       <div className={styles.sliderCard}>
         <img className={styles.sliderCardImg} src={src} alt={title} />
         <h3 className={styles.sliderCardTitle}>{title}</h3>
@@ -19,9 +19,8 @@ function SliderCard({ title, desc, price, src, link }) {
 SliderCard.propTypes = {
   title: PropTypes.string.isRequired,
   desc: PropTypes.string.isRequired,
-  price: PropTypes.string.isRequired,
+  price: PropTypes.number.isRequired,
   src: PropTypes.string.isRequired,
-  link: PropTypes.string.isRequired,
 };
 
 export default SliderCard;
