+++ src/routes/Item/Item.module.css
@@ -1,68 +1,108 @@
-.productImg{
-  width: 90uubupx; 
+.container{
+  font-family: Raleway, sans-serif;
+  height: 85%;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+
 }
 
-.productImgSection{
-  display: flex; 
+.productImgSection {
+  display: flex;
   justify-content: center;
   margin-bottom: 1rem;
 }
 
-
-.container{
-  font-family: Raleway, sans-serif;
+.productImg {
+  width: 120px;
+  margin-top: -60px;
 }
 
-.listItem {
-  font-size: var(--13px);
-  line-height: var(--20px);
+.button{
+  color:var(--primary-color);
+  margin: 0.5rem 0; 
+  border: none;
+
 }
 
 
 
-.footerNavItem {
-  cursor: pointer;
-  fill: transparent;
-  stroke: #200e32;
-}
 
-.footerNavItemActive {
-  composes: footerNavItem;
-  filter: drop-shadow(0px 0px 15px rgba(89, 86, 233, 1));
-  fill: var(--primary-color);
-  stroke: var(--primary-color);
-}
+
 
 /* 360px */
 @media only screen and (min-width: 22.5rem) {
+  .productImg {
+    width: 150px;
+    margin-top: -20px;
+  }
 
- 
+  .button{
+    font-size: 15px;
+    margin: 1rem 0; 
+  }
 }
 
 /* 480px */
 @media only screen and (min-width: 30rem) {
- 
+  .productImg {
+    width: 200px;
+    margin-top: 20px; 
+  }
+
+  .button{
+    font-size: 14px;
+  }
 }
 
 /* 600px */
 @media only screen and (min-width: 37.5rem) {
+  .productImg {
+    width: 240px;
+    margin-top: 30px; 
+  }
+  .button{
+    font-size: 14px;
+  }
 }
 
 /* 736px */
 @media only screen and (min-width: 46em) {
- 
+  .productImg {
+    width: 260px;
+    margin-top: 30px; 
+  }
+  .button{
+    font-size: 15px;
+  }
 }
 
 /* 980px */
 @media only screen and (min-width: 61.25rem) {
- 
+  .productImg {
+    width: 360px;
+  }
+  .button{
+    font-size: 22px;
+  }
 }
 
 /* 1280px */
 @media only screen and (min-width: 80rem) {
-  
+  .productImg {
+    width: 460px;
+  }
+  .button{
+    font-size: 22px;
+  }
 }
 
 /* 1680 px */
 @media only screen and (min-width: 105rem) {
+  .productImg {
+    width: 490px;
+  }
+  .button{
+    font-size: 23px;
+  }
 }
