+++ src/routes/Item/index.jsx
@@ -7,6 +7,7 @@ import Header from "../../components/Header";
 import TotalPrice from "../../components/TotalPrice";
 import ProductInfo from "./productinfo";
 import ArrowIcon from "../../../public/assets/arrow-icon.svg";
+// import { useItems } from "../../contexts/ItemsContext";
 
 const { container, productImg, productImgSection } = styles;
 
@@ -17,22 +18,27 @@ const product = {
   longDesc:
     "Available when you purchase any new iPhone, iPad, iPod Touch, Mac or Apple TV, 4.99/month after free trial.",
   fullDesc: "",
-  color: ["Sky Blue", "Rose Gold", "green"],
+  color: [
+    ["Sky Blue", "#80BDE3"],
+    ["Rose Gold", "#B76E79"],
+    ["green", "#9FE2BF"],
+  ],
   price: 359,
   src: "assets/apple-watch-red.png",
 };
 
 function Item() {
+  // const items = useItems();
+
   return (
     <Container white>
+      <Header icon="heart" iconColor="black" />
       <div className={container}>
-        <Header icon="heart" iconColor="black" />
-
         <section className={productImgSection}>
           <img className={productImg} src={product.src} alt={product.title} />
         </section>
 
-        <section className="productInformation">
+        <section>
           <ProductInfo product={product} />
 
           <TotalPrice amount={product.price} />
@@ -40,11 +46,10 @@ function Item() {
           <button type="button" className={styles.button}>
             Full Description <ArrowIcon />
           </button>
-
-          <Link to="/home">
-            <Button inverted>Add to basket</Button>
-          </Link>
         </section>
+        <Link to="/home">
+          <Button inverted>Add to basket</Button>
+        </Link>
       </div>
     </Container>
   );
